requirements.txt
python-magic>=0.4.18
pillow>=7.2.0
cryptography>=2.9.2
selenium>=3.141.0
six>=1.15.0
typing>=3.7.4.1
numpy>=1.19.0
python-telegram-bot>=12.8